+++ CHANGES.rst
@@ -3,6 +3,7 @@
 
 - Remove test dependency on astropy. [#4]
 
+- Force garbage collection before checking for open files. [#7]
 
 0.1 (2017-10-09)
 ================